+++ src/diff-ignore/diff-ignore.js
@@ -46,7 +46,8 @@ export async function execute(node) {
 
     // Update the "Files changed" summary header to reflect the removed diffs count
     const summaryHeader = node.querySelector(
-        '#pullrequest-diff > section.main > h1, #commit-summary > h1, #compare-diff-content > h1'
+        '#pullrequest-diff > section.main > h1, #commit-summary > h1, #compare-diff-content > h1, ' +
+            '#diff > h1'
     )
     summaryHeader.textContent += ` - Showing ${filesChanged.length -
         filesToRemove.length} of ${filesChanged.length}`